requirements.txt
# Phase 1 — Story, Script & Character Design
# Core dependencies

# LLM / Agent framework
groq>=0.9.0
langgraph>=0.2.0
langchain>=0.3.0

# Schema validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-mock>=3.14.0

# Utilities
httpx>=0.27.0
requests
Pillow
moviepy
huggingface-hub
ffmpeg-python
imageio-ffmpeg